Refactor cameraView
